services/scraper_difarmer/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.1
google-generativeai>=0.3.0
selenium>=4.10.0
webdriver-manager>=4.0.0
requests>=2.31.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
regex>=2023.6.3
python-multipart
twilio
